# Buff later naqfuels and misc changes