--- a/src/main/java/goodgenerator/loader/FuelRecipeLoader.java
+++ b/src/main/java/goodgenerator/loader/FuelRecipeLoader.java
@@ -6,6 +6,7 @@ import static goodgenerator.main.GG_Config_Loader.NaquadahFuelTime;
 import static goodgenerator.main.GG_Config_Loader.NaquadahFuelVoltage;
 import static gregtech.api.enums.Mods.GTPlusPlus;
 import static gregtech.api.util.GT_RecipeBuilder.SECONDS;
+import static gregtech.api.util.GT_RecipeBuilder.TICKS;
 import static gregtech.api.util.GT_RecipeConstants.LNG_BASIC_OUTPUT;
 import static gregtech.api.util.GT_RecipeConstants.NFR_COIL_TIER;
 
@@ -111,7 +112,7 @@ public class FuelRecipeLoader {
                 MyMaterial.orundum.get(OrePrefixes.dust, 64))
             .fluidInputs(
                 MyMaterial.naquadahBasedFuelMkIII.getFluidOrGas(2000),
-                new FluidStack(FluidRegistry.getFluid("molten.hypogen"), 1440))
+                new FluidStack(FluidRegistry.getFluid("molten.hypogen"), 720))
             .fluidOutputs(MyMaterial.naquadahBasedFuelMkIV.getFluidOrGas(500))
             .duration(8 * SECONDS)
             .eut(75_000_000)
@@ -119,17 +120,36 @@ public class FuelRecipeLoader {
             .noOptimize()
             .addTo(naquadahFuelRefineFactoryRecipes);
 
+        // One-step recipe to allow easier scaling for MK VI
+        GT_Values.RA.stdBuilder()
+            .itemInputs(
+                MyMaterial.extremelyUnstableNaquadah.get(OrePrefixes.dust, 54),
+                MyMaterial.orundum.get(OrePrefixes.dust, 32),
+                ItemRefer.High_Density_Uranium.get(10),
+                ItemRefer.High_Density_Plutonium.get(5))
+            .fluidInputs(
+                MyMaterial.heavyNaquadahFuel.getFluidOrGas(4000),
+                MyMaterial.lightNaquadahFuel.getFluidOrGas(5000),
+                new FluidStack(FluidRegistry.getFluid("molten.hypogen"), 360),
+                new FluidStack(FluidRegistry.getFluid("molten.chromaticglass"), 6480))
+            .fluidOutputs(MyMaterial.naquadahBasedFuelMkIV.getFluidOrGas(250))
+            .duration(10 * TICKS)
+            .eut(350_000_000)
+            .metadata(NFR_COIL_TIER, 4)
+            .noOptimize()
+            .addTo(naquadahFuelRefineFactoryRecipes);
+
         // MK V Naquadah Fuel
         GT_Values.RA.stdBuilder()
             .itemInputs(
-                GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Infinity, 16),
+                GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Infinity, 8),
                 MyMaterial.atomicSeparationCatalyst.get(OrePrefixes.dust, 32))
             .fluidInputs(
                 MyMaterial.naquadahBasedFuelMkIV.getFluidOrGas(2000),
-                FluidRegistry.getFluidStack("heavyradox", 1000))
+                FluidRegistry.getFluidStack("heavyradox", 250))
             .fluidOutputs(MyMaterial.naquadahBasedFuelMkV.getFluidOrGas(500))
-            .duration(100_000_000)
-            .eut(10 * SECONDS)
+            .duration(10 * SECONDS)
+            .eut(100_000_000)
             .metadata(NFR_COIL_TIER, 2)
             .noOptimize()
             .addTo(naquadahFuelRefineFactoryRecipes);
@@ -137,12 +157,12 @@ public class FuelRecipeLoader {
         // Alternate higher tier recipe
         GT_Values.RA.stdBuilder()
             .itemInputs(
-                GT_OreDictUnificator.get(OrePrefixes.dust, MaterialsUEVplus.SpaceTime, 8),
+                GT_OreDictUnificator.get(OrePrefixes.dust, MaterialsUEVplus.SpaceTime, 4),
                 GT_OreDictUnificator.get(OrePrefixes.dust, MaterialsUEVplus.TranscendentMetal, 16),
                 MyMaterial.atomicSeparationCatalyst.get(OrePrefixes.dust, 48))
             .fluidInputs(
                 MyMaterial.naquadahBasedFuelMkIV.getFluidOrGas(2000),
-                FluidRegistry.getFluidStack("heavyradox", 1000))
+                FluidRegistry.getFluidStack("heavyradox", 250))
             .fluidOutputs(MyMaterial.naquadahBasedFuelMkV.getFluidOrGas(750))
             .duration(10 * SECONDS)
             .eut(300_000_000)
@@ -157,7 +177,7 @@ public class FuelRecipeLoader {
                 GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Tritanium, 32))
             .fluidInputs(
                 MyMaterial.naquadahBasedFuelMkV.getFluidOrGas(2000),
-                FluidRegistry.getFluidStack("molten.shirabon", 1440))
+                FluidRegistry.getFluidStack("molten.shirabon", 360))
             .fluidOutputs(MyMaterial.naquadahBasedFuelMkVI.getFluidOrGas(500))
             .duration(12 * SECONDS)
             .eut(320_000_000)
@@ -168,14 +188,14 @@ public class FuelRecipeLoader {
         // Alternate higher tier recipe
         GT_Values.RA.stdBuilder()
             .itemInputs(
-                GT_OreDictUnificator.get(OrePrefixes.dust, MaterialsUEVplus.WhiteDwarfMatter, 8),
+                GT_OreDictUnificator.get(OrePrefixes.dust, MaterialsUEVplus.WhiteDwarfMatter, 4),
                 ELEMENT.STANDALONE.CELESTIAL_TUNGSTEN.getDust(64),
                 GT_OreDictUnificator.get(OrePrefixes.dust, Materials.Tritanium, 48))
             .fluidInputs(
                 MyMaterial.naquadahBasedFuelMkV.getFluidOrGas(2000),
                 FluidRegistry.getFluidStack("molten.shirabon", 1440))
             .fluidOutputs(MyMaterial.naquadahBasedFuelMkVI.getFluidOrGas(750))
-            .duration(12 * SECONDS)
+            .duration(4 * SECONDS)
             .eut(530_000_000)
             .metadata(NFR_COIL_TIER, 4)
             .noOptimize()