**Naquadah Reactor tooltip:**
- Add info that Tachyon Rich Temporal Fluid only consumes 20L/s instead of 1000L/s

![image](https://github.com/user-attachments/assets/d1dcb5ed-2e4a-4a5d-920d-4008b93fea0f)


**Refinery tooltip:**
- Change "adjoin" -> "adjacent to" in structure tooltip

**Mk4 Naquadah Fuel:**
Alternate recipe:
- Change Hypogen consumption from 1440L to 720L

Added new recipe to one-step Mk4 fuel, reducing Mk3+Mk4 Refinery spam with T4 coils
Recipe:
- 54 Extremely Unstable Naquadah Dust
- 32 Orundum Dust
- 10 High Density Uranium
- 5 High Density Plutonium
- 4000L Heavy Naquadah Fuel
- 5000L Light Naquadah Fuel
- 360L Molten Hypogen
- 6480L (45 dust ) Molten Chromatic Glass

![image](https://github.com/user-attachments/assets/c464f086-c3fd-40b0-b463-b3d56c2d27da)


Produces 250L at 10 ticks per recipe.
The recipe combines the Tier 3 and Tier 4 recipes, while being around 4 times faster. This should reduce the refinery spam for scaling up in UMV.
It uses 4 fluids because the refinery does not accept more than 4 item inputs in the recipes (??? Couldn't figure out why)

**Mk5 Naquadah Fuel:**
Primary Recipe
- Infinity Dust 16 -> 8
- Heavy Radox 1000L -> 250L
- Time from 5 000 000s to 10s (RA2 Fix)
- Usage from 200 EU/t to 100 000 000 EU/t (RA2 Fix)

Alternate Recipe:
- Spacetime Dust 8 -> 4
- Heavy Radox from 1000L to 250L
- Time from 10s to 5s


**Mk6 Naquadah Fuel:**
Primary Recipe
- Molten Shirabon 1440L -> 360L

Alternate Recipe:
- Molten Shirabon 1440L -> 360L
- Time from 12s to 8s
- White Dwarf Matter 8 -> 4

**Additional changes:**
- Changes the default NEI recipe display widget to support 4 fluids in the default view better.
- Change references from "Naquadah Fuel Refine Factory" -> "Naquadah Fuel Refinery"

Closes https://github.com/GTNewHorizons/GT-New-Horizons-Modpack/issues/16425